requirements.txt
aiohttp==3.10.3
Flask-Migrate==4.0.7
annotated-types==0.6.0
arrow==1.3.0
asyncio==3.4.3
datefinder==0.7.3
dateparser==1.2.0
docker==7.1.0
exceptiongroup==1.2.0
Flask==3.0.3
flask-cors==4.0.0
flask-smorest==0.44.0
flask-sqlalchemy==3.1.1
h2==4.1.0
httptools==0.6.1
importlib-metadata==7.1.0
ipykernel==6.29.3
marshmallow-sqlalchemy==1.0.0
pickleshare==0.7.5
pip-chill==1.0.3
#psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic-core==2.16.3
pytest==8.1.1
python-dotenv==1.0.1
pyyaml==6.0.1
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.16
supabase==2.6.0
sutime==1.0.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
